requirements.txt
# Smart Caffeine Analytics Dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Airtable integration
pyairtable>=2.0.0
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Date handling
python-dateutil>=2.8.0

# Async support
aiohttp>=3.8.0

# Logging
structlog>=23.1.0